moved all configuration type files to config dir on server side
