+++ src/server/config/passport.js
@@ -1,13 +1,18 @@
 const JwtStrategy = require('passport-jwt').Strategy;
-const ExtractJwt = require('passport-jwt').ExtractJwt;
-const {User} = require('./users/userModel');
-const {SECRET} = require('./config');
+const {User} = require('../users/userModel');
+const {SECRET} = require('./app.config');
 
-//passport strategy that extracts JWT from header and gets user id
+// Extract JWT cookie if it exists
+const extractJwtCookie = function(req) {
+  let token = null;
+  if (req && req.cookies) token = req.cookies['token'];
+  return token;
+};
 
+// Passport strat that pulls jwt from req cookie
 module.exports = function(passport){
   const opts = {};
-  opts.jwtFromRequest = ExtractJwt.fromAuthHeader();
+  opts.jwtFromRequest = extractJwtCookie;
   opts.secretOrKey = SECRET;
   passport.use(new JwtStrategy(opts, function(jwt_payload, done) {
     User
@@ -20,7 +25,6 @@ module.exports = function(passport){
             done(null, user);
           } else {
             done(null, false);
-          }
-      });
+          }});
   }));
-}
+};
